Fixes #517 - show progress for crc downloads

+++ framework/bundles/org.jboss.tools.rsp.runtime.core/src/main/java/org/jboss/tools/rsp/runtime/core/model/IDownloadRuntimeConnectionFactory.java
@@ -14,5 +14,5 @@ import java.net.URL;
 
 public interface IDownloadRuntimeConnectionFactory {
 	public InputStream createConnection(URL url, String user, String pass);
-	public int getContentLength(URL url, String user, String pass);
+	public long getContentLength(URL url, String user, String pass);
 }
